Major Revisions
Required data is now included.
Glitched data removed.
Errors in configuration file fixed.
Paths are now relative.
Selection components for the on- and off-phase analyses have been improved.

Explanations and examples added for if the user doesn't want to use the pre-processed data. This section also includes a tutorial on using the bash tools to add the PULSE_PHASE column to the evfile.
Added explanations on why certain choices are being made.